Bind forward trampolines to a symbol by default (#8)
* Bind trampolines to symbol by default.

This is a major rewrite of our trampoline generator so just the emitters first and

* Testing future 'bound' and classic 'unbound' trampolines.

This is a quick hack to see if tests and fuzzing work. I'll flip the default trampoline system if this passes.

* Still not happy with this but I'm done for today.

* Plan for errors

* Start using the new error system.

I'll need to write tests but that'll come later...

* Minor fixes to serialization's sake.

* Attempt to trick clang on ARM while running coverage tests

See https://github.com/sanko/infix/actions/runs/18317837036/job/52162808854

* Nevermind.

* Looking into moving to xmake over lengthy perl script